refactor(tools): keep instructions with adapters
