+++ src/tool-surfaces/codex.ts
@@ -5,6 +5,7 @@ import type { ProcessSnapshot } from "../process-sessions.js";
 import {
   EDIT_TOOL_ANNOTATIONS,
   SHELL_TOOL_ANNOTATIONS,
+  toolNames,
   workspaceIdDescription,
   type ToolRegistrationContext,
 } from "./types.js";
@@ -18,6 +19,12 @@ import {
 } from "./shared.js";
 
 type CodexRegistration = (context: ToolRegistrationContext) => void;
+
+const CODEX_INSTRUCTIONS = `Use ${toolNames.read} for direct file reads, apply_patch for all file modifications, exec_command for inspection, tests, builds, and other commands, and write_stdin to poll or interact with running processes. Follow instructions returned by ${toolNames.openWorkspace}; read applicable instruction and skill files before working in their scope.`;
+
+export function codexInstructions(): string {
+  return CODEX_INSTRUCTIONS;
+}
 
 export function registerCodexTools(context: ToolRegistrationContext): void {
   for (const register of CODEX_REGISTRATIONS) {
